Temporarily update crtbegin_dynamic.o in order to resolve circular dependency
Update crtbegin_dynamic.o with one built locally with the rest of the
topic in order to resolve a circular dependency between landing the
bionic change that updates crtbegin_dynamic, landing the Soong change that
needs the new prebuilt, and updating the prebuilts.  Will be replaced
with an official version after the topic lands.

Test: treehugger
Change-Id: I8de72f290215918aa059ab12aa1cb3cd43a940fc