Drop exported_bootclasspath_fragments from prebuilt com.android.xxx APEXes.
This avoids ambiguity when deciding which prebuilt APEX to use for
deapexing in platform builds where there are other prebuilts than
these.

This is a workaround - a proper solution for this requires introducing
a product variable to specify which APEXes to use, but that requires a
lot of engineering effort (b/192006406).

exported_bootclasspath_fragments is only required to use these prebuilt
APEXes in system images, and currently there are no plans to use these
there - they are only used in chroot testing on the thin manifest
master-art branch.

Test: m droid SOONG_CONFIG_art_module_source_build=false
  on internal master
Test: art/build/build-art-module.sh
  on master-art
Test: art/tools/buildbot-build.sh --target
  on master-art
Bug: 192006406
Bug: 202192894
Change-Id: I6ec7047e3fae90b19a6216b1022fe0faac2b1222