Add annotations zip, and system and module-lib APIs
At the moment they are all empty but they are needed to ensure that
the prebuilts provide the same outputs as the source modules.

The module-lib and system APIs are empty in the sense that they provide
no additional APIs on top of the public API. Their stub sources are
not empty but are in fact copies of the public API sources.

Updated the files by first running this command:
   m i18n-module-sdk

Then unzip the generated snapshot into a temporary directory. Then copy
the differences in the Android.bp file related to the annotations and
APIs from that directory into this. Also copy the annotations.zip
files. Finally, copied the public API stubs source into module-lib and
system.

BYPASS_INCLUSIVE_LANGUAGE_REASON=copying existing API files

Bug: 216435117
Test: m nothing
      - TH will test the prebuilts in master-art
Change-Id: Iec2e23041e19db090a1a1de2b5f30d982952453b