Update prebuilt drop script for new CI build locations.
Split InstallSdkEntries into the three variants required for the
different types of SDKs produced by
packages/modules/common/build/mainline_modules_sdks.py.

InstallApexEntries and InstallSharedLibEntries are also refactored
slightly, but there are currently no CI builds to support them.

From now on we want to be conservative with prebuilt updates to ensure
SC compatibility in these dependencies, so update the script and
instructions to reflect this. Remove the --install-modules filter since
it isn't granular enough, and instead comment out most modules which
may not be safe to update.

Test: prebuilts/runtime/mainline/update.py -b aosp-master --build 8376291
      banchan com.android.art && m
      art/build/apex/runtests.sh
Bug: 216762037
Change-Id: I95573df27bc238240d8f12acf1c313b4473be6a3