Enable art/tools/ahat to use libicu_jni.
`ahat-tests` requires the `libicu_jni` library at runtime. This change
allows ahat-tests to use them.

This is the `prebuilt` version of `libicu_jni`, but it takes a long
time for the source code version to be officially pushed to the
prebuilt version. Since adding visibility is safe in itself, we will
update the visibility of `libicu_jni` together with the source code
version for now.

Bug: 342330305
Test: m -j libicu_jni; atest ahat-tests
Change-Id: I38f2fdd32d93c81cb0e72c69e56482ebb7652d9d